requirements.txt
fastapi
uvicorn
torch
Pillow
torchvision
python-multipart